# laa-pdf-ai-hackathon-api/generated/src/main/java/uk/gov/justice/laa/springboot/microservice/api/ImageApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uk.gov.justice.laa.springboot.microservice.api;

import uk.gov.justice.laa.springboot.microservice.model.ImageResponse;
import uk.gov.justice.laa.springboot.microservice.model.ImageSummary;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-16T13:27:37.508336+01:00[Europe/London]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "image", description = "the image API")
public interface ImageApi {

    String PATH_GET_IMAGE_BY_ID = "/api/v1/image/{id}";
    /**
     * GET /api/v1/image/{id} : Get image submission by ID
     *
     * @param id The ID of the image submission to retrieve (required)
     * @return Success (status code 200)
     *         or Not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getImageById",
        summary = "Get image submission by ID",
        tags = { "image" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ImageResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ImageApi.PATH_GET_IMAGE_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<ImageResponse> getImageById(
        @NotNull @Parameter(name = "id", description = "The ID of the image submission to retrieve", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    String PATH_LIST_IMAGES = "/api/v1/image";
    /**
     * GET /api/v1/image : List all image submissions
     *
     * @return Success (status code 200)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "listImages",
        summary = "List all image submissions",
        tags = { "image" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ImageSummary.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ImageApi.PATH_LIST_IMAGES,
        produces = { "application/json" }
    )
    ResponseEntity<List<ImageSummary>> listImages(
        
    );


    String PATH_UPLOAD_IMAGE = "/api/v1/image";
    /**
     * POST /api/v1/image : Upload an image for processing
     *
     * @param image The image file to be processed (required)
     * @param email Email address of the submitter (required)
     * @return Created (status code 201)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "uploadImage",
        summary = "Upload an image for processing",
        tags = { "image" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ImageResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ImageApi.PATH_UPLOAD_IMAGE,
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    ResponseEntity<ImageResponse> uploadImage(
        @Parameter(name = "image", description = "The image file to be processed", required = true) @RequestPart(value = "image", required = true) MultipartFile image,
        @Parameter(name = "email", description = "Email address of the submitter", required = true) @jakarta.validation.constraints.Email @Valid @RequestParam(value = "email", required = true) String email
    );

}