# laa-pdf-ai-hackathon-api/generated/src/main/java/uk/gov/justice/laa/springboot/microservice/api/ItemsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uk.gov.justice.laa.springboot.microservice.api;

import uk.gov.justice.laa.springboot.microservice.model.Item;
import uk.gov.justice.laa.springboot.microservice.model.ItemRequestBody;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-16T13:27:37.508336+01:00[Europe/London]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "items", description = "the items API")
public interface ItemsApi {

    String PATH_CREATE_ITEM = "/api/v1/items";
    /**
     * POST /api/v1/items : Create an item
     *
     * @param itemRequestBody create an item (optional)
     * @return Created (status code 201)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createItem",
        summary = "Create an item",
        tags = { "items" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created"),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ItemsApi.PATH_CREATE_ITEM,
        consumes = { "application/json" }
    )
    ResponseEntity<Void> createItem(
        @Parameter(name = "ItemRequestBody", description = "create an item") @Valid @RequestBody(required = false) @Nullable ItemRequestBody itemRequestBody
    );


    String PATH_DELETE_ITEM = "/api/v1/items/{id}";
    /**
     * DELETE /api/v1/items/{id} : Delete item by id
     *
     * @param id The id of the item to be deleted (required)
     * @return No content (status code 204)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "deleteItem",
        summary = "Delete item by id",
        tags = { "items" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No content"),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ItemsApi.PATH_DELETE_ITEM
    )
    ResponseEntity<Void> deleteItem(
        @NotNull @Parameter(name = "id", description = "The id of the item to be deleted", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    String PATH_GET_ITEM_BY_ID = "/api/v1/items/{id}";
    /**
     * GET /api/v1/items/{id} : Get item by id
     *
     * @param id The id of the item to retrieve (required)
     * @return Success (status code 200)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getItemById",
        summary = "Get item by id",
        tags = { "items" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Item.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ItemsApi.PATH_GET_ITEM_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<Item> getItemById(
        @NotNull @Parameter(name = "id", description = "The id of the item to retrieve", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    String PATH_GET_ITEMS = "/api/v1/items";
    /**
     * GET /api/v1/items : Get items
     *
     * @return Success (status code 200)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getItems",
        summary = "Get items",
        tags = { "items" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Item.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ItemsApi.PATH_GET_ITEMS,
        produces = { "application/json" }
    )
    ResponseEntity<List<Item>> getItems(
        
    );


    String PATH_UPDATE_ITEM = "/api/v1/items/{id}";
    /**
     * PUT /api/v1/items/{id} : Update an item
     *
     * @param id The id of the item to be updated (required)
     * @param itemRequestBody update an existing item (optional)
     * @return No content (status code 204)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "updateItem",
        summary = "Update an item",
        tags = { "items" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No content"),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ItemsApi.PATH_UPDATE_ITEM,
        consumes = { "application/json" }
    )
    ResponseEntity<Void> updateItem(
        @NotNull @Parameter(name = "id", description = "The id of the item to be updated", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "ItemRequestBody", description = "update an existing item") @Valid @RequestBody(required = false) @Nullable ItemRequestBody itemRequestBody
    );

}