# laa-pdf-ai-hackathon-api/open-api-specification.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Items API
paths:
  /api/v1/items:
    get:
      tags:
        - items
      summary: Get items
      operationId: getItems
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Items"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
    post:
      tags:
        - items
      summary: Create an item
      operationId: createItem
      requestBody:
        description: create an item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestBody'
      responses:
        '201':
          description: 'Created'
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

  /api/v1/items/{id}:
    get:
      summary: Get item by id
      operationId: getItemById
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the item to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
    put:
      tags:
        - items
      summary: Update an item
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the item to be updated
          schema:
            type: integer
            format: int64
      requestBody:
        description: update an existing item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestBody'
      responses:
        '204':
          description: No content
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
    delete:
      summary: Delete item by id
      operationId: deleteItem
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the item to be deleted
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No content
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

  /api/v1/image:
    get:
      tags:
        - image
      summary: List all image submissions
      operationId: listImages
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageSummary'
        '500':
          description: 'Internal server error'
    post:
      tags:
        - image
      summary: Upload an image for processing
      operationId: uploadImage
      requestBody:
        description: Multipart form containing an image file and associated metadata
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageRequestBody'
      responses:
        '201':
          description: 'Created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '500':
          description: 'Internal server error'

  /api/v1/image/{id}:
    get:
      tags:
        - image
      summary: Get image submission by ID
      operationId: getImageById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the image submission to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

components:
  schemas:
    Item:
      required:
        - id
        - name
        - description
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    ItemRequestBody:
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string

    Items:
      type: array
      items:
        $ref: "#/components/schemas/Item"

    ImageSummary:
      required:
        - id
        - email
        - submittedAt
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the submission
        email:
          type: string
          description: Email address of the submitter
        submittedAt:
          type: string
          format: date-time
          description: Timestamp when the submission was received

    ImageRequestBody:
      required:
        - image
        - email
      properties:
        image:
          type: string
          format: binary
          description: The image file to be processed
        email:
          type: string
          format: email
          description: Email address of the submitter

    ImageResponse:
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the created image submission
        extractedData:
          type: object
          description: Structured JSON data extracted from the image by the AI provider
          additionalProperties: true